Fix executor/TEST106
This change adds synchronization between the sqlci running
the subject query and the sqlci which cancels that query
for the escalation test case and the update statistics test
case. It also logs and tests the ordering of events recorded
in T106_interleaved_log, to make it easier to debug the test.
Finally, it increases the timeout interval in the
synchronization helper script to 1 hour.

Change-Id: I26926bfb3831ccc1245be2e230f395148ade57e0