ODB by Maurizio Felici
ODB is a platform independent, multi-threaded,
ODBC command line tool you can use as:
-          data loader/extractor
-          query driver
-          SQL interpreter

A manual is being prepared and a draft is available upon request.
ODB has been included in the Trafodion build process and will be
available as part of the Trafodion client package. It will also
be available in $MY_SQROOT/export/bin64. Using the tool requires
configuration of the ODBC environment which is described in the
manual. There are future plans to automate this for the standard
development environment and cluster installs.

In this release a 64 bit version that runs on Linux and uses
unixODBC driver manager is available. Versions of the clients
for other platforms and driver managers will be available in
later releases.

The Trafodion team welcomes and thanks Maurizio for his valuable
contribution.

Change-Id: I596324c5cf33fab67b466fbe430bec2e43476c1b