Writing packed explain plan to metric_query_table
The packed explain is retrieved using the SQL_EXEC_GetExplainData() CLI
call. This is picked up by the repository watchdog thread and then written to
the metric_query_table using SQL_EXEC_StoreExplainData().

The packed explain could have binary data within it so it could not be
used as is to formulate the insert statement the way we do today with a
stringstream. So, Anoop provided the new CLI call which takes in the plan
and updates the row in the metric_query_table table.

Also, removed writing of sql text and plan during the updates to
metric_query_table since they don't change and is redundant.

Change-Id: I15efc2134b78d0d43000b1b05094aface4d9d0e0