Eliminate manual steps in load/ustat integration
The fix achieves full integration of the bulk load utility with
Update Statistics. The Hive backing sample table is now creeated
automatically (formerly, we only wrote the HDFS files to be
used by the Hive external table), the correct sampling percentage
for the sample table is calculated, and the ustat command is
launched fro1m the executor as one of the steps in execution of
the bulk load utility.

Change-Id: I9d5600c65f0752cbc7386b1c78cd10a091903015
Closes-Bug: #1436939