Changes to reduce the memory growth/leak in mxosrvr and T2 driver
Memory growth in mxosrvr
It was observed that the memory growth comes from JNIHandle::alloc_block
within JVM in mxosrvr.  PushLocalFrame and PopLocalFrame is now called
in Trafodian JNI methods calling java functions to ensure this memory
growth is avoided.

CommonLogger was allocating and deallocating buffer to log the message
every time. Instead a thread safe variable is used to allocate only
once per thread.

CommonLogger was constructing the message even when the category is NOT
set. The message construction involves string creation and memory allocation.

Changes in T2.
The buffer to hold the output row was allocated every time. It is now
allocated once and reused till the statement is dropped.

The row count buffer was allocated but was used after deallocation.

Change-Id: I7aad93beb03dc42a85b78da720737e2eb46a7080