Adding break out of preSplit() loop on exception
Also commenting out log message that is filling up client-side
logs even if there is no activity.  The zookeeper message will
still be printed as DEBUG to show if the ZNode is not present.

The break of out preSplit() was needed as on a RS abort, the
preSplit can get stuck in a state where it is keeping the RS
process up and in a while/sleep loop.

Change-Id: Ic93ded73d80ecb61c6ceda94339dc53f52ea1bf1