ClientSide batch RegionServer 2-phase commit calls
Disabled by default, use the option in ms.env to enable:
TM_BATCH_REGIONSERVER=1

Included in this change are the server-side stubs which
are needed for building as new protocol buffers have
been added to support the multiple regionserver sets
of names. Server-side changes will be added later.

Basic design: On the client-side we have a list of
all the regions we need to communicate with.  On prepare/
commit/abort we send those calls out individually by region.
This change will group the region names per regionserver
into one protocol buffer and include that in the RPC
call to the server.

Error handling is done by collecting the problematic
regions into a separate list in the TransactionState.
These 'retry' regions are then retried using the
single-region calls.

Implemented coprocessor calls using single-region method.

Change-Id: I968f061ff603d55956c6b541a9bda1968e57481f