Bug #1451618 MXOSRVR returns trunctated string to clients
When a table has varchar column and the varchar column size
is defined bigger than 32K, MXOSRVR server may return truncated
string to client application in certain circumstances.

Fix in sql/cli/CliExpExchange.cpp to not clear trailing characters.
As per SQL team this code is not needed and is a left over.

Also fixed alignment issues as pointed out by SQL and Connectivity
team (related to change done for bug #1449343)

This alignment change in mxosrvr needs a corresponding change in
linux and windows odbc driver. Windows driver change will be
submitted as a separate checkin.

Change-Id: Iabde0d3a0ae2922e15dbdcd36d23ba367da967fb