DDL Tx, Changes to handle upsert using load operations on abort
SQL statements 'upsert using load' will be executed outside of a
transaction even though the operation can performed inside a
transaction. For this reason, it is necessary for us to handle this
data differently when a transaction aborts or fails.

These operations will be registered as part of the TmDDL manager
and if the transaction is aborted, that table gets truncated. The data
loaded will be deleted but table will still exist. At the moment, this is
handled by using disable, delete and create hbase methods. However, when
hbase is upgraded to the next version, we will use the truncate hbase
method option.

Change-Id: Ica9d81a284b5f1821a3048b9c8deaad449a4c4f4