Option to use ConcurrentHashMap vs synchronizedSet
Default is synchronizedSet

Option is set in ms.env
DTM_USE_CONCURRENTHM=1

The synchronization option is for the participatingRegions
list in the client-side TransactionState.

We are seeing an issue where a duplicate region is being
inserted into the particpant list and it is possible that a
synchronizedSet is not blocking concurrent access as we
expected.

Change-Id: I433c0adc513f6a5336d88f6373d8f296cbc28edd