Part 1 of ALTER TABLE/INDEX ALTER HBASE_OPTIONS support
blueprint alter-table-hbase-options

This set of changes includes the following:
1. Syntax for ALTER TABLE/INDEX ALTER HBASE_OPTIONS
2. Compiler binding and generation logic
3. Run-time logic to update the metadata TEXT table with the
new options.

Still to come is:
1. Logic to actually change the HBase object
2. Transactional logic to actually change the HBase object

The functionality in this set of changes is only marginally
useful. If you manually change hbase options on a Trafodion
object using the hbase shell, you could use this ALTER
TABLE/INDEX command to update the Trafodion metadata. (Of
course some care would have to be taken to use the same
options!).

This work is part of

Change-Id: Id0a5513fe80853c06acdbbf6cc9fd50492fd07b2