Adding support for incremental index maintenance during bulk load.
Previously when bulk loading into a table with indexes, the indexes are first
disabled, base table is loaded and then the indexes are populated from
scratch one by one. This could take a long time when the table has significant
data prior to the load.

Using a design by Hans this change allows indexes to be loaded in the same
query tree as the base table. The query tree looks like this

          Root
           |
       NestedJoin
        /      \
      Sort    Traf_load_prep (into index1)
       |
     Exchange
       |
   NestedJoin
   /       \
 Sort     Traf_load_prep (i.e. bulk insert) (into base table)
  |
Exchange
  |
Hive scan

This design and change set allows multiple indexes to be on the same tree.
Only one index is shown here for simplicity. LOAD CLEANUP and LOAD COMPLETE
statements also now perform these tasks for the base table along with all
enabled indexes

This change is enabled by default. If a table has indexes it will be
incrementally maintained during bulk load.
The WITH NO POPULATE INDEX option has been removed
A new option WITH REBUILD INDEXES has been added. With this option we get
the old behaviour of disabling all indexes before load into table and
then populate all of them from scratch.

Change-Id: Ib5491649e753b81e573d96dfe438c2cf8481ceca