Repository support in MT server #1
This is the first installment of code to start including repository
functionality in the DCS multi-threaded server/T2. Includes only session statistics code for now.

Note: This code is for compile only for now and not part of any active
code path.

There are some temporary to-do comments in the code which shall be removed eventually.

Change-Id: Ida3689d9c0f029ebb47e6d87e670d0a4a1653480