Removed about 50 obsolete make files from sql
Security is not built inside sql, and sqlutils is not in Trafodion, so
their makefiles are not needed and have been deleted.  Also deleted a few
other unused makefiles.  In modified makefiles, removed unused parts.

The sql/nskgmake/Makerules* files still have references to sqlutils, those
lines will be considered in a separate change.

Change-Id: Id9ab492d4a23579cd0ab6db7279e0da56c55ea4c